subiendo prueba

+++ ft_strlcat.c
@@ -6,20 +6,20 @@
 /*   By: samusanc <samusanc@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 16:25:43 by samusanc          #+#    #+#             */
-/*   Updated: 2023/01/17 16:25:44 by samusanc         ###   ########.fr       */
+/*   Updated: 2023/02/09 15:35:48 by samusanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include<stdio.h>
 #include "libft.h"
 
-static void	ft_chng_str(int i, int i3, const char *s, char *d)
+static void	ft_chng_str(size_t i, size_t i3, const char *s, char *d)
 {
-	int		i4;
-	int		i2;
+	size_t	i4;
+	size_t	i2;
 
-	i2 = (int)(ft_strlen(d) + ft_strlen((char *)s));
-	i4 = i2 - (int)ft_strlen((char *)s);
-	while (i4 < (int)i - 1 && i3 < (int)(ft_strlen((char *)s)))
+	i2 = ft_strlen(d) + ft_strlen((char *)s);
+	i4 = i2 - ft_strlen((char *)s);
+	while (i4 < i - 1 && i3 < ft_strlen((char *)s))
 	{
 		d[i4] = s[i3];
 		++i4;
@@ -30,14 +30,14 @@ static void	ft_chng_str(int i, int i3, const char *s, char *d)
 
 size_t	ft_strlcat(char *d, const char *src, size_t i)
 {
-	char	*s;
-	int		i2;
-	int		i3;
+	char		*s;
+	size_t		i2;
+	size_t		i3;
 
 	s = (char *)src;
 	if (!i)
 		return (ft_strlen(s));
-	i2 = (int)(ft_strlen(d) + ft_strlen(s));
+	i2 = ft_strlen(d) + ft_strlen(s);
 	i3 = 0;
 	if (i > ft_strlen(d))
 	{
